GG-36231 / remove excessive casting

+++ modules/control-utility/src/test/java/org/apache/ignite/util/GridCommandHandlerPropertiesTest.java
@@ -233,7 +233,7 @@ public class GridCommandHandlerPropertiesTest extends GridCommandHandlerClusterB
         Method nextCheckpointIntervalMtd = Checkpointer.class.getDeclaredMethod("nextCheckpointInterval");
         nextCheckpointIntervalMtd.setAccessible(true);
 
-        return (long) (Long) nextCheckpointIntervalMtd.invoke(database.getCheckpointer());
+        return (Long)nextCheckpointIntervalMtd.invoke(database.getCheckpointer());
     }
 
     /**
